refactor: rename backend.routes to api.routes
