+++ server/src/routes/users.routes.ts
@@ -1,5 +1,5 @@
 import { Hono } from 'hono';
-import { Env, Vars } from '../backend.routes';
+import { Env, Vars } from '../api.routes';
 import { handleApiError, handleApiSuccess } from '../utils/api.utils';
 import { CreateUserData, User } from '../types';
 import { UserService } from '../services/user.service';
